+++ services/aztec-listener/src/svcs/poller/pollers/block_poller/index.ts
@@ -1,30 +1,24 @@
-import { AztecAddress, L2Block } from "@aztec/aztec.js";
 import { ChicmozL2BlockFinalizationStatus } from "@chicmoz-pkg/types";
 import {
   AZTEC_DISABLE_LISTEN_FOR_PROPOSED_BLOCKS,
   AZTEC_DISABLE_LISTEN_FOR_PROVEN_BLOCKS,
   BLOCK_POLL_INTERVAL_MS,
   CATCHUP_POLL_WAIT_TIME_MS,
-} from "../../../environment.js";
-import { onBlock, onCatchupBlock } from "../../../events/emitted/index.js";
-import { logger } from "../../../logger.js";
+} from "../../../../environment.js";
+import { onBlock, onCatchupBlock } from "../../../../events/emitted/index.js";
+import { logger } from "../../../../logger.js";
 import {
   getBlockHeights,
   storeBlockHeights,
   storeProcessedProposedBlockHeight,
   storeProcessedProvenBlockHeight,
-} from "../../database/heights.controller.js";
+} from "../../../database/heights.controller.js";
 import {
-  deletePendingTx,
-  getAllPendingTxs,
-} from "../../database/pending-txs.controller.js";
-import { publishMessage } from "../../message-bus/index.js";
-import {
-  getBalanceOf,
   getBlock,
   getLatestProposedHeight,
   getLatestProvenHeight,
-} from "../network-client/index.js";
+} from "../../network-client/index.js";
+import { handleProvenTransactions } from "./handle-proven-block.js";
 
 let timeoutId: number | undefined;
 let cancelPolling = false;
@@ -156,7 +150,6 @@ const pollProposedBlock = async (height: number, isCatchup: boolean) => {
 const pollProvenBlock = async (height: number, isCatchup: boolean) => {
   const block = await internalGetBlock(height);
 
-  // Handle proven transactions
   await handleProvenTransactions(block);
 
   if (isCatchup) {
@@ -170,65 +163,6 @@ const pollProvenBlock = async (height: number, isCatchup: boolean) => {
     await onBlock(block, ChicmozL2BlockFinalizationStatus.L2_NODE_SEEN_PROVEN);
   }
   await storeProcessedProvenBlockHeight(height);
-};
-
-const handleProvenTransactions = async (block: L2Block) => {
-  try {
-    // Get all pending txs from DB
-    const pendingTxs = await getAllPendingTxs();
-    if (pendingTxs.length === 0) {
-      return;
-    }
-
-    // Get transaction hashes from the proven block
-    const provenTxHashes = block.body.txEffects.map((txEffect) => {
-      const hash = txEffect.txHash;
-      return hash.toString();
-    });
-
-    // Find pending txs that are now proven
-    const provenPendingTxs = pendingTxs.filter((pendingTx) =>
-      provenTxHashes.includes(pendingTx.txHash),
-    );
-
-    if (provenPendingTxs.length === 0) {
-      return;
-    }
-
-    const blockNumber = Number(block.header.globalVariables.blockNumber);
-    logger.info(
-      `🎯 Found ${provenPendingTxs.length} proven pending txs in block ${blockNumber}`,
-    );
-
-    // For each proven pending tx, query balance and publish event
-    for (const provenTx of provenPendingTxs) {
-      try {
-        // Convert feePayer string to AztecAddress
-        const feePayerAddress = AztecAddress.fromString(provenTx.feePayer);
-
-        // Query balance of feePayer
-        const balance = await getBalanceOf("latest", feePayerAddress);
-
-        // Publish contractInstanceBalance event
-        await publishMessage("CONTRACT_INSTANCE_BALANCE_EVENT", {
-          contractAddress: provenTx.feePayer,
-          balance: balance.toString(),
-          timestamp: new Date(),
-        });
-
-        // Remove pending tx from DB
-        await deletePendingTx(provenTx.txHash);
-
-        logger.info(
-          `✅ Processed proven tx ${provenTx.txHash}, balance: ${balance.toString()}`,
-        );
-      } catch (error) {
-        logger.error(`Error processing proven tx ${provenTx.txHash}:`, error);
-      }
-    }
-  } catch (error) {
-    logger.error("Error handling proven transactions:", error);
-  }
 };
 
 const internalGetBlock = async (height: number) => {
